requirements.txt
albumentations==1.1.0
h5py==2.10.0
matplotlib==3.3.4
matplotlib-inline==0.1.3
notebook==6.4.4
numba==0.55.1
numpy==1.21.6
opencv-python==4.5.5.64
opencv-python-headless==4.1.2.30
pytorch-lightning==1.5.10
pytorch-ranger==0.1.1
scikit-build==0.12.0
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
segmentation-models-pytorch @ git+https://github.com/qubvel/segmentation_models.pytorch@8bf52c7e862af006e76a23aae6aa17977d7f9a79
-e git+ssh://git@github.com/sithu31296/semantic-segmentation.git@958ed542aa68003eb0a2b0799cf5cecfe6c7587c#egg=semseg
six==1.16.0
spacy==3.1.3
spacy-legacy==3.0.8
pandas==1.1.5
Pillow==8.1.1
plotly==5.7.0
tabulate==0.8.9
tf-estimator-nightly==2.8.0.dev2021122109
torch==1.8.1
torch-optimizer==0.3.0
torch-summary==1.4.5
torchaudio==0.9.0
torchmetrics==0.7.2
torchvision==0.10.0+cu111
tqdm==4.62.3
wandb==0.12.7